+++ src/pages/services/services-routers.ts
@@ -1,11 +1,11 @@
-import Introduction from './introduction';
+import Introduction from './introduction/introduction';
 import QuarterlyChart from './ndvi-temp/index';
 import Page3 from './page3';
 import Page4 from './page4';
 import Page5 from './page5';
 import Overview from './overview/index';
-import RSImagery from './rs-imagery';
-import Xylophilus from './xylophilus';
+import RSImagery from './rs-imagery/rs-imagery';
+import Xylophilus from './xylophilus/xylophilus';
 
 export {
     Introduction,
